requirements.txt
google-generativeai>=0.3.0
python-dotenv==1.0.0
faster-whisper==1.2.0
gTTS==2.5.4
pygame==2.6.0
